requirements.txt
feedparser>=6.0.0
requests>=2.31.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
Pillow>=10.0.0
